pyproject.toml
[project]
name = "stream-snip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pytchat>=0.5.5",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
]
